+++ controllers/users.controller.js
@@ -1,23 +1,38 @@
-const config = require("../config/db.config.js")
+const jwt = require("jsonwebtoken");
+const bcrypt = require("bcryptjs");
+
+const config = require("../config/db.config.js");
 const db = require("../models/index.js");
 const User = db.user;
+const Favorite = db.favorites
 
 const { Op, ValidationError, Sequelize } = require("sequelize");
 
 // Obtains general information about all users. Route only available for admins (authentication token must be provided in header). Has an optional limit counter.
 exports.findAll = async (req, res) => {
-  let { page, limit } = req.query;
+  let { page, limit, sort } = req.query;
 
   const pageNumber = page && Number.parseInt(page) > 0 ? Number.parseInt(page) : 1;
-  // const sizeNumber = limit && Number.parseInt(limit) > 0 ? Number.parseInt(limit) : 10;
-
-  // const limitValue = sizeNumber;
   const limitValue = limit && Number.parseInt(limit) > 0 ? Number.parseInt(limit) : 10;
   const offset = (pageNumber - 1) * limitValue;
 
+  if (sort) {
+    if (sort.toLowerCase() != 'asc' && sort.toLowerCase() != 'desc') {
+      return res.status(400).json({
+        success: false,
+        message: "Sort can only be 'asc' or 'desc'."
+      });
+    }
+  }
 
   try {
-    let users = await User.findAll({ limit: limitValue, offset: offset, raw: true });
+    if (req.loggedUserRole !== "admin")
+      return res.status(403).json({
+        success: false,
+        msg: "You don't have permission to access this route."
+      });
+
+    let users = await User.findAll({ limit: limitValue, offset: offset, order: [['username', sort.toUpperCase()]], raw: true });
 
     users.forEach((user) => {
       user.links = [
@@ -58,10 +73,25 @@ exports.findAll = async (req, res) => {
 // Handles user registration to join the platform
 exports.register = async (req, res) => {
   try {
-    const emailFound = await User.findOne({ where: { email: req.body.email } });
+    if (!req.body && !req.body.username && !req.body.password)
+      return res.status(400).json({ success: false, msg: "Username and password are mandatory" });
 
-    if (!emailFound) {
-      let newUser = await User.create(req.body);
+    let searchUser = await User.findOne({ where: { username: req.body.username } })
+    if (searchUser) {
+      res.status(409).json({
+        success: false,
+        msg: "The username is already taken. Please choose another one."
+      });
+    } else {
+      let newUser = await User.create({
+        username: req.body.username, email: req.body.email,
+        password: bcrypt.hashSync(req.body.password, 10),
+        phone_number: req.body.phone_number, profile_image: req.body.profile_image,
+        first_name: req.body.first_name, last_name: req.body.last_name,
+        is_confirmed: req.body.is_confirmed, user_role: req.body.user_role,
+        owner_description: req.body.owner_description
+      });
+
       res.status(201).json({
         success: true,
         msg: "User created successfully.",
@@ -70,12 +100,7 @@ exports.register = async (req, res) => {
           { rel: "login-user", href: `/users/login`, method: "POST" }
         ],
       });
-    } else {
-      res.status(409).json({
-        msg: "The email address is already associated with another account.",
-      });
     }
-
   } catch (error) {
     if (error instanceof ValidationError) {
       res
@@ -97,73 +122,109 @@ exports.register = async (req, res) => {
 
 // Obtains information about specified user. Route only available for admins (authentication token must be provided in header).
 exports.findUser = async (req, res) => {
+  let { field } = req.query;
+
   try {
     let userFound = await User.findByPk(req.params.idT)
-    console.log(userFound.user_role);
-    let user
-    if(userFound.user_role == 'owner'){
-      user = await User.findByPk(req.params.idT, {
-        include: [
-          {
-            model: db.favorites,
-            as: 'favorites',
-            attributes: ['property_ID']
-          },
-          {
-            model: db.property,
-            as: 'properties',
-            attributes: ['title']
-          },
-          {
-            model: db.message,
-            as: 'properties',
-            attributes: ['content']
-          },
-          {
-            model: db.message,
-            as: 'messages_received',
-            attributes: ['content']
-          },
-        ]
-      });
-    }else{
-      user = await User.findByPk(req.params.idT, {
-        include: [
-          {
-            model: db.favorites,
-            as: 'favorites',
-            attributes: ['property_ID']
-          },
-          {
-            model: db.reservation,
-            as: 'reservations',
-            attributes: ['dateIn']
-          },
-          {
-            model: db.review,
-            as: 'reviews',
-            attributes: ['comment']
-          },
-          {
-            model: db.message,
-            as: 'messages_sent',
-            attributes: ['content']
-          },
-          {
-            model: db.message,
-            as: 'messages_received',
-            attributes: ['content']
-          },
-        ]
+
+    if (!userFound) {
+      return res.status(404).json({
+        success: false,
+        msg: "The specified username does not exist.",
       });
     }
 
+    let user
 
-    if (user === null) {
-      return res.status(404).json({
-        success: false,
-        msg: `The specified username does not exist.`,
-      });
+    if (field) {
+      if (field == 'favorites') {
+        user = await User.findByPk(req.params.idT, {
+          include: [
+            {
+              model: db.favorites,
+              as: 'favorites',
+              attributes: ['property_ID']
+            }
+          ]
+        });
+      } else if (field == 'reservations') {
+        user = await User.findByPk(req.params.idT, {
+          include: [
+            {
+              model: db.reservation,
+              as: 'reservations',
+              attributes: ['dateIn']
+            }
+          ]
+        });
+      } else if (field == 'properties') {
+        user = await User.findByPk(req.params.idT, {
+          include: [
+            {
+              model: db.property,
+              as: 'properties',
+              attributes: ['title']
+            }
+          ]
+        });
+      }
+    } else {
+      if (userFound.user_role == 'owner') {
+        user = await User.findByPk(req.params.idT, {
+          include: [
+            {
+              model: db.favorites,
+              as: 'favorites',
+              attributes: ['property_ID']
+            },
+            {
+              model: db.property,
+              as: 'properties',
+              attributes: ['title']
+            },
+            {
+              model: db.message,
+              as: 'messages_sent',
+              attributes: ['content']
+            },
+            {
+              model: db.message,
+              as: 'messages_received',
+              attributes: ['content']
+            },
+          ]
+        });
+      } else {
+        user = await User.findByPk(req.params.idT, {
+          include: [
+            {
+              model: db.favorites,
+              as: 'favorites',
+              attributes: ['property_ID']
+            },
+            {
+              model: db.reservation,
+              as: 'reservations',
+              attributes: ['dateIn']
+            },
+            {
+              model: db.review,
+              as: 'reviews',
+              attributes: ['comment']
+            },
+            {
+              model: db.message,
+              as: 'messages_sent',
+              attributes: ['content']
+            },
+            {
+              model: db.message,
+              as: 'messages_received',
+              attributes: ['content']
+            },
+          ]
+        });
+      }
     }
 
     return res.json({
@@ -197,7 +258,7 @@ exports.findUser = async (req, res) => {
   }
 };
 
-// Update a user
+
 exports.update = async (req, res) => {
   try {
     let user = await User.findByPk(req.params.idT);
@@ -242,21 +303,43 @@ exports.update = async (req, res) => {
   }
 };
 
-// Delete one user
+// Handles user account’s deletion.
 exports.delete = async (req, res) => {
   try {
-    let result = await User.destroy({ where: { username: req.params.idT } });
-    if (result == 1) {
-      return res.json({
-        success: true,
-        msg: `User permanently deleted successfully.`,
-      });
-    }
+    if (req.loggedUserRole === "admin") {
+      let result = await User.destroy({ where: { username: req.params.idT } });
+      if (result == 1) {
+        return res.json({
+          success: true,
+          msg: `User permanently deleted successfully.`,
+        });
+      }
 
-    return res.status(404).json({
-      success: false,
-      msg: `The specified username does not exist.`,
-    });
+      return res.status(404).json({
+        success: false,
+        msg: `The specified username does not exist.`,
+      });
+    } else {
+      if (req.loggedUserId == req.params.idT) {
+        let result = await User.destroy({ where: { username: req.params.idT } });
+        if (result == 1) {
+          return res.json({
+            success: true,
+            msg: `Your account has been permanently deleted.`,
+          });
+        }
+
+        return res.status(404).json({
+          success: false,
+          msg: `Your account does not exist.`,
+        });
+      } else {
+        return res.status(403).json({
+          success: false,
+          msg: `You are not authorized to delete other users.`,
+        });
+      }
+    }
   } catch (error) {
     if (error instanceof Sequelize.ConnectionError) {
       res.status(503).json({
@@ -275,7 +358,21 @@ exports.delete = async (req, res) => {
 // Handles user login.
 exports.login = async (req, res) => {
   try {
-    return res.status(200).json({ success: true })
+    if (!req.body || !req.body.username || !req.body.password)
+      return res.status(400).json({ success: false, msg: "Must provide username and password." });
+
+    let user = await User.findOne({ where: { username: req.body.username } });
+    if (!user) return res.status(404).json({ success: false, msg: "User not found." });
+    const check = bcrypt.compareSync(req.body.password, user.password);
+    if (!check) return res.status(401).json({ success: false, accessToken: null, msg: "Invalid credentials!" });
+
+    const token = jwt.sign({ id: user.username, role: user.user_role },
+      config.SECRET, {
+      expiresIn: '24h' // 24 hours
+    });
+
+    return res.status(200).json({ success: true, accessToken: token });
+
   } catch (error) {
     if (error instanceof Sequelize.ConnectionError) {
       res.status(503).json({
@@ -291,7 +388,6 @@ exports.login = async (req, res) => {
   }
 };
 
-// Handle user recovery email on POST
 exports.recoverEmail = async (req, res) => {
   try {
     console.log('Here');
@@ -314,11 +410,56 @@ exports.recoverEmail = async (req, res) => {
 };
 
 // Handle user adding favorites on POST
-exports.favorites = async (req, res) => {
+exports.addFavorite = async (req, res) => {
   try {
-    console.log('Here');
-    return res.json({
-      "message": "Property added to favorites."
+    if (!req.body.property_ID)
+      return res.status(400).json({
+        success: false,
+        msg: "Property ID is mandatory"
+      });
+
+    await Favorite.create({
+      username: req.params.idT,
+      property_ID: req.body.property_ID,
+    });
+
+    res.status(200).json({
+      success: true,
+      msg: "Property added to favorites.",
+      links: [
+        { rel: "self", href: `/users/${req.params.idT}?field=favorites`, method: "GET" }
+      ],
+    });
+
+  } catch (error) {
+    if (error instanceof Sequelize.ConnectionError) {
+      res.status(503).json({
+        error: "Database Error",
+        msg: "There was an issue connecting to the database. Please try again later"
+      });
+    } else {
+      res.status(500).json({
+        error: "Server Error",
+        msg: "An unexpected error occurred. Please try again later."
+      });
+    }
+  }
+};
+
+// Removes specified property from favorites 
+exports.removeFavorite = async (req, res) => {
+  try {
+    let result = await Favorite.destroy({ where: { property_ID: req.params.idP } });
+    if (result == 1) {
+      return res.json({
+        success: true,
+        msg: `Property removed from favorites.`,
+      });
+    }
+
+    return res.status(404).json({
+      success: false,
+      msg: `The specified property isn't favorited.`,
     });
   } catch (error) {
     if (error instanceof Sequelize.ConnectionError) {
